Remove CS+ attacks from list result.
If one would set a pubkey A as the CS+ name for account B, a search for exactly B would return only account of A.
This would also happen when searching through QRcode.

This commit has been tested on a web instance, not on a smartphone build. Make sure it works.